% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecture_data.R
\name{delete_variables}
\alias{delete_variables}
\title{delete_variables}
\usage{
delete_variables(data, threshold = 0.1)
}
\arguments{
\item{data}{data table object}

\item{threshold}{delete variable with a percent of missing value > 1 - threshold}
}
\value{
a data.table object
}
\description{
delete variables under the threshold
}
