% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecture_data.R
\name{read_anticho}
\alias{read_anticho}
\title{read_anticho}
\usage{
read_anticho(
  file = "CH-Rouffach-Prescriptions-2008-2018_v1.txt.csv",
  sep = "|",
  dec = ".",
  header = TRUE,
  quote = "",
  fill = FALSE
)
}
\arguments{
\item{file}{file name}

\item{sep}{separator}

\item{dec}{decimal symbol}

\item{header}{true or false}

\item{quote}{quote parameter}

\item{fill}{true or false (fill the non available data by NA)}
}
\value{
a data.table object (see data.table package for further information)
}
\description{
Reading a csv file of prescription data
}
