% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lecture_data.R
\name{standardize_variables}
\alias{standardize_variables}
\title{standardize_variables}
\usage{
standardize_variables(data)
}
\arguments{
\item{data}{data table object}
}
\value{
standardized data.table object
}
\description{
changing variable type, process missing data
}
